requirements.txt
requests
cryptography
PyNaCl
